Refactor StaticSelectElement
